Phil, this is the sort of thing I'm talking about.
Note I'd like to use the standard header :

/**
 * \file cvsx4IHSg
 * Copyright 2002 OProfile authors
 * Read the file COPYING
 *
 * \author John Levon <moz@compsoc.man.ac.uk>
 */

(add \author where appropriate) since it's miles shorter
and more informative.

Pros of splitting things up :

1. Depdencies of smaller scope. Now nothing else at all needs
   to know cplus_demangle() goop.

2. Simple source navigation

3. Interface simply presented by the .h file

Cons :

1. more complicated build (you know one day I will make us use automake ...)

2. more source files

This example is fairly extreme (short span of code), but pollution with DMGL_PARAMS etc.
in other files is now gone.

And think if we want to say work around a demangler bug or something