Use a structure for the module sample buffer.
Alex/Bob, this changes the interface for you. Now you read in a structure :

	struct op_buffer_head {
		int cpu_nr; // the cpu number of the buffer
		int count; // number of entries in buffer
		struct op_sample buffer[0];
	}

from the read of the oprofile device and the buffer is placed at the end here.
Simplifies some things, and allows cpu-specific profiles to happen. You must
read exactly sizeof(op_buffer_head) + sizeof(op_sample) * bufsize bytes. If
no samples are present, then a buffer head is still copied in, with a count of 0.