Symptom:  If runtime binaries have been stripped of symbol information, users must make debuginfo files available that contain the symbol information for those bianries so that opreport can attribute samples to symbols for those binaries.  But in a prelinked environment (e.g., RHEL, Fedora), opreport does not correctly process the debuginfo files, thus generating report data such as the following:
samples  %        image name               symbol name
1531      6.2205  libcairo.so.2.11000.2    /usr/lib64/libcairo.so.2.11000.2

In the example given above, the debuginfo package corresponding to the cairo
package has been installed.  Note that this problem only occurs where the
runtime binary is completely stripped of symbol information.  So on RHEL and
Fedora, basic  symbol information is retained in the runtime libc library, so
this symptom does not appear with samples taken in libc.

This patch reverts previous changes that actually broke this functionality.
The changes being reverted are:
  - The SEC_LOAD filter
  - Calling translate_debuginfo_syms for all architectures