Discard user context kernel samples for which we have no app_cookie
Symptom: Under certain conditions, samples are not attributed to the appropriate
app name.  If profiling is done with --separate=lib,kernel,thread and then a
report is generated using a tgid or tid specification to see profile data for one
particular process or thread, you may see anomalies like that shown below.  Note
the line with "ext3" under the 'app name' column, when all other lines (rightly)
have "spin".

 samples  %        image name               app name                 symbol name
 1018088  99.8455  spin                     spin                     .spin
 1566      0.1536  ext3                     spin                     /ext3
 4        3.9e-04  ext3                     ext3                     /ext3
 1        9.8e-05  ld-2.12.1.so             spin                     ._dl_map_object_from_fd
 1        9.8e-05  ld-2.12.1.so             spin                     ._dl_relocate_object
 1        9.8e-05  ld-2.12.1.so             spin                     .check_match.8837
 1        9.8e-05  libc-2.12.1.so           spin                     ._IO_cleanup

See the big comment in daemon/opd_sfile.c for details of the problem and how
this patch fixes it.