Fix opannotate --assembly to work with prelinked runtime lib and non-prelink debuginfo
The opannotate tool can annotate assembly code as long as the binary file
contains symbol information.  For binaries that are completely stripped,
users should install the corresponding debuginfo file and then use
'opannotate --assembly' to obtain annotated assembly output for the binary.
However, in a prelink environment, 'opannotate --assembly' does not produce
any output for stripped binaries, even when it's associated debuginfo file
is installed.  This patch fixes that problem by making a 'vma_adj' value
available to the opannotate code.  The 'vma_adj' value is set to either:
1) the difference between the runtime load address of the binary and the p_vaddr
of the first PT_LOAD segment of the debuginfo file; or 2) 0, when the runtime
binary *does* have symbol info.  The reason for #2 is that if the real binary does
have symbol info, then the VMA values of the samples will already be adjusted
according to the runtime load address of the symbol from which the sample was
taken, prior to the annotation steps.