Add support for --pid option and fix timing bug with passing COMMAND option.
A bug was reported when using operf as follows:
   operf -r ping 92.22.100.1
The first invocation of this worked OK, but subsequent invocations failed with
  "perf_event_open failed with Permission denied"

This was due to an error in how the code was handling the exec of the passed
COMMAND (i.e., program to be profiled).  We needed to add synchronization so
that the exec did not happen before the perf_event_open call.  As I was looking
into how to fix this, it seemed like a good time to also add support for the --pid
option, so I did.