Add support for kernel profiling; incidental namespace fixups
The bulk of this patch is for adding support to operf for recording
kernel samples.  The vmlinux option can be used to pass in a vmlinux
file to use for symbol resolution and debuginfo (if available in
that file).  If no vmlinux option is passed in, a default "no-vmlinux"
pseudo image is created to which all kernel samples are attributed.

This patch also includes some namespace fixups that were incidental
to the kernel profiling enhancment. Additionally, a few small changes
were made in the way a passed application name is processed.