Fix timing-related problems with app name
Under certain circumstances when doing system-wide profiling, operf
may received MMAP events for depenedent libraries before receiving
the corresponding COMM event which represents an executable being
profiled.  The COMM event only contains a short "app name" of 16
characters, so we need to try match up a COMM event with an MMAP
event to obtain our best guess for the app pathname.  There were
a couple of bugs in the area of the code that was supposed to
handle the case described above (where MMAP's come before COMM).
This patch fixes those bugs.  I also renamed the "app_name" field
of the operf_process_info class so it doesn't collide visually
with a global of the same name.  This was not actually causing a
probelem since the usage spaces of the two variables were completely
separate, but it just caused some confusion in understanding the
code.