Re-process samples dropped due to no mapping or no process info
The perf_events kernel subsystem often starts recording sample
events before recording corresponding COMM and MMAP events.
So in the first pass through the recorded events, it's not
unusual to fail to match up a sample with its process info or
the binary mapping (executable or library) for which the sample
was taken.  With this patch, we save such samples; then, after
the first pass through all recorded events is complete, we
re-process those saved samples to try again to match them up with
their corresponding process info and mapping.