Make new --separate-thread option for operf
This patch reverses the earlier decision to do "separate=thread" type
of functionality by default.  So now there is a command line option
to operf called "--separate-thread".  I decided to make this an option
for two reasons: 1) operf run in system-wide mode required the user to
either pass a tgid spec or to use "--merge=tgid" or the report was
unreadable;  2) typically, the Java JIT compiler will spawn threads
to execute JITed code; so when using operf to profile a single Java app,
the resulting report is usually unreadable without using a tid spec or
"--merge=tid".