Change operf to create oprofile sample files during profiling instead of afterwards
Previously, the 'operf-record' process read the perf_events sample data from the kernel
and wrote it to a temporary output file (operf.data).  Then, once the profiling session
was completed, a conversion was performed on that data to put it into oprofile format,
stored in the usual sample files in <session-dir>/samples/current.  I noticed that when
operf was run in system-wide mode, the temporary operf.data file was very large, and the
conversion step was taking inordinately long (from several seconds on a laptop to over
a minute on a large, busy server system.

I've had a TODO comment in the operf.cpp code to investigate performing the conversion
step in-flight (while profiling) versus waiting for the profiling to end.  So I changed
the operf-record process from writing the perf_events sample data to a file to writing
to a pipe.  Then I created a new operf-read process to read from the other end of that
pipe and perform the conversion to oprofile format.  Doing the conversion during the
profiling run allowed me to identify the hot spots in the opreport output. I found several
areas (especially involving the inefficiency of the operf_sfile hash algorithm) where
improvements were needed.  The conversion step now takes about 15% of the time it used
to! A wonderful result!