Flesh out operf support for kernel versus user domain sample recording.
Added support for handling event specifications of the form:

   -e <event_name>:<count>:<unit_mask_num>:<kernel_mode>:<user_mode>

The '<kernel_mode>' and '<user_mode>' parts can be either '1' or '0',
indicating to either record samples or not, respectively, for the given
mode.

This patch also fixes an unrelated problem where samples were sometimes
be mis-attributed due to having an "operf_transient" object in an
inconsistent state.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>