Factor out /proc/cpuinfo scanning and add support for ARM to operf
This patch adds support for parsing the cpu type from /proc/cpuinfo on
ARM platforms and hooks this into __get_cpu_type_alt_method so that
perf-events can be used as the backend. The previous /proc/cpuinfo parsing
code used by ppc64 to identify the current CPU type is useful for other
architectures that don't have a userspace cpuid instruction (for example, ARM),
so this patch factors out the /proc/cpuinfo reading code to be usable for
ARM support.

Signed-off-by: Will Deacon <will.deacon@arm.com>