Fix opjitconv to not end abnormally when no jit dump directory is found
When operf is run on a system before any other opcontrol commands,
it's likely that the jit dump directory (/var/lib/oprofile/jitdump)
will not exist, since that directory is created by opcontrol.
However, the opjitconv program (which is run at the end of operf
execution) expects to find this directory and will end abnormally
if it does not find it.  This patch allows opjitconv to end normally
in either case.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>