Handle EINTR of operf read of sample data pipe
When profiling an application with operf's "--pid" option,
the user must stop profiling using ctrl-C.  A signal
handler was setup for the operf process that reads
sample data from a pipe (written to by the operf record
process), but when that handler was triggered, the
operf read process was ending before reading the
rest of the sample data from the pipe.  In cases
where lots of sample data has already been received,
the little bit that gets left in the pipe may not be
noticeable.  But for situations where the samples
are few and far between, this problem can, on some
platforms, result in no profiling data being received
at all.

The fix for the problem is to detect an EINTR when trying
to read the pipe and go back and read it again.

Another symptom of unnecessary stopping due to EINTR is
possible when opening the sample data files to store
the converted sample data.  This is a very small window,
and I have not actually seen this symptom, but with this
patch, I plug up that window as well.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>