Update support for ARMv7 processors
  - update architected ARMv7 event descriptions
  - add support for Cortex-A5 ARMv7 PMU implementations
  - add support for Cortex-A7 ARMv7 PMU implementations
  - add support for Cortex-A15 ARMv7 PMU implementations

The part of this patch that adds support for new ARM PMU
implementations also updates the new function in
libop/op_cpu_type that was added in the perf-events branch
so that these new processors can be used with operf.
For this reason, this patch is currently only being applied
to the perf-events branch, so support for the new ARM processors
won't be available in oprofile "mainline" until we merge the
perf-events branch with the master branch.

Signed-off-by: Will Deacon <will.deacon@arm.com>