Fix error when using 'operf --callgraph --events=blah'
When executing the following command:
   'operf -g -e CPU_CLK_UNHALTED:500000 ls'

the following error message occurs:
   "Invalid count for event --callgraph=1"

I tracked this down to the switch-over I made on June 11 to
change operf from using popt to use getopt_long instead.  The
error message is coming from where operf invokes the ophelp
command as follows:
   'ophelp --check-events CPU_CLK_UNHALTED:500000 --callgraph=1'

This ophelp command works fine if you execute it from the command
line.  It also works fine when executed by opcontrol.  So the only
situation it doesn't work correctly is when executed by operf.
I put some debug printf's into ophelp and found that when it calls
poptGetArgs(), the returned value is '2' in the above example when
operf is used, but is '1' when ophelp is called from the command line
or opcontrol.  My initial thought was that either operf
or ophelp might be making improper use of some memory, but running
both programs under Valgrind revealed no problems. So now,
my theory is that because popt uses getopt under the
covers, there must be some global variable being used in operf's use
of getopt that later calls something to go awry when ophelp uses
the poptGetArgs function.

The fix for this problem is to have operf pass the "--callgraph=1"
option to ophelp before passing the event spec.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>