Allow libpfm4 to be used by ppc64 for obtaining event code
The ppc64 architecture oprofile event codes are not the same
as are needed for perf_events, so there is ppc64-specific
code to use libpfm to obtain the correct code to pass to
perf_event_open.  The libpfm API has undergone a lot of
changes from version 3 to version 4, and some of the version 3
functions have been removed, replaced, deprecated, etc.
This patch detects which version of libpfm is installed
and #define's some macros so that the ppc64-specific code
can be built with the appropriate functions.

Note that at this time, this patch affects *only* the
ppc64 architecture.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>