Change pp tools to abort if <cur_dir>/oprofile_data exists, but no samples found
In commit 01c7ce945cfe835b908c5f4f8c1125e65302b000 on July 3, 2012,
a change was made to the post-processing tools to look first in
<cur_dir>/oprofile_data; then, if no samples are found there,
the tool should fall back to the "standard" /var/lib/oprofile
session-dir.  I've come the conclusion that this is not proper
behavior.  For example, a user may run operf using some event that
never occurs during the profiling run, and, thus, no samples would
be stored in <cur_dir>/oprofile_data.  But if that user then ran
opreport and there happened to be some old profile data in
/var/lib/oprofile, that old profile data would be displayed in
the report.

This new patch changes the behavior as follows:
  The post-processing analysis tools will search for samples in
  <current_dir>/oprofile_data first. If that directory exists
  but has no samples, then the usual "No sample file found"
  message is displayed.  On the other hand, if that directory
  does not exist, the post-processing tools use the standard
  session-dir of /var/lib/oprofile.  Also, the message that
  indicates the session-dir being used will be displayed
  whether or not samples are found so that users can see
  clearly where the pp tool was looking for samples.

This patch also makes a small change in the "No sample file found"
message relating to the use of 'opcontrol --dump', which is only
appropriate when using legacy profiling versus operf profiling.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>