operf: print lost sample warning only if lost samples exceeds threshold
Currently a warning is printed if there are any lost samples when collecting
data using the operf tool.  This patch changes the code to print the warning
when the total number of lost samples, for whatever reason, exceeds a
threshold of the total number of samples.  Initially, the threshold has been
set to 0.01% of the total number of samples.

Signed-off-by: Carl Love <cel@us.ibm.com>