Fix opimport to not try to import .jo ELF files
If JIT profiling was done on the target system (e.g., passing
'-agentlib:jvmti_oprofile' to the JVM), then a synthesized ELF
file named "<pid>.jo" will be created for each JVM that was profiled,
and this file will be stored in the samples directory. However, we
do no want opimport to treat such files as sample data files, because
it may corrupt them and make them unreadable for when opreport tries
to get symbol information from it.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>