Oprofile, operf:  Expanded error message when running multiple operf sessions
As a non-root user, you can only run one session of operf at a time.  The
current error message does not fully explain all the possible causes for operf
not being able to run.  This patch adds an additional explanation as to the
possible cause for operf not being able to run.

Signed-off-by: Carl Love <cel@us.ibm.com>