Fix opjitconv to handle process IDs up to PID_MAX_LIMIT
When using the libjvmti_oprofile.so java agent library to profile
a Java app, a <pid>.dump file is created that stores information
about the JITed code. The opjitconv program is supposed to use that
<pid>.dump file to create an ELF-like file named <pid>.jo, which is
the binary file against which samples are attributed by opreport.
But if the process ID is greater than the Linux default of 32768,
opjitconv fails to create the <pid>.jo, so all samples are attributed
to the java anon memory mappings, and there is no visibility to the
java methods (symbols) that were JITed.

Current PID_MAX_LIMIT (as defined in include/linux/threads.h) is
   4 x 4 x 1024 * 1024 (for 64-bit kernels)
This patch changes opjitconv to handle 7 chars for proc_id.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>