Handle early perf_events kernel without PERF_RECORD_MISC_GUEST* macros
In very early versions of perf_events kernel subsystem, the
PERF_RECORD_MISC_GUEST_KERNEL and PERF_RECORD_MISC_GUEST_USER
macros (in perf_event.h) were not yet defined.  This patch adds
a configure check to determine when it's OK for source code to refer
to those macros.

This patch also does some minor cleanup of the configure script
help and warning messages relating to the --with-kernel option.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>