Fix bug in finding command in PATH
If operf has to search PATH to obtain the full pathname of the
passed app (or command) and a segment in PATH does not exist,
the following error will occur:

        <cmd> cannot be found in your PATH.

The workaround is to specify the full pathname of the command
when passing it to operf or remove the non-existent segment in
the PATH environment variable.

This patch fixes this problem in operf.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>