Remove daemon/liblegacy since 2.4 kernels are no longer supported
A build problem reported by Vineet Gupta (Vineet.Gupta1@synopsys.com)
involving a compile failure of liblegacy code when built under uClibc
made it clear that we missed some things when support for 2.4 kernels
was removed in Oct 2012.  The daemon/liblegacy directory is all about
support for 2.4 kernels, so this patch completely removes that directory.
Of course, this caused some breakage elsewhere in the oprofile tree
which needed fixing up.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>