Oprofile operf:  Fix the code to strip the _GRP## from the event name
The current code uses the call "strstr(event.name, "_GRP")" to find the
substring for the group number at the end of the POWER events.  The
strstr() function finds the first occurance of the substring processing
from left to right.  This will find the string "_GRP_" in the name of
the event rather then the intended _GRP## at the end of the string.  For
example the event name "PM_GRP_CMPL_GRP174" is currently change to "PM"
instead of "PM_GRP_CMPL".  This patch makes a change in the calculation
for the strncpy() call to use the function rfind("_GRP") to return the index
of where the last instance of the substring is found.  Basically the call
finds the first occurance of the substring by searching from right to
left.

Signed-off-by: Carl Love <cel@us.ibm.com>