Add support for using build-id to locate debuginfo files
This patch addresses the issue raised in oprofile bug #3591165.
In brief, a new MiniDebuginfo feature was introduced in Fedora18 that
changes how runtime RPMs and their corresponding debuginfo RPMs are built
such that the CRC stored in the runtime binary no longer matches the CRC
we calculate from the contents of the debuginfo file. Other tools that use
debuginfo (like gdb) use the the build-id method (i.e., build-id is stored
in runtime which can be used to locate the matching debuginfo file), and
only use the CRC method as a fallback mechanism. This patch adds the
build-id method, making it the primary mechanism for locating debuginfo
files.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>