Fix opreport header info on unit mask when operf is run without a UM specified
When a user runs operf and profiles with an event that needs a unit mask value,
the default unit mask value will be used if no UM value is specified.  When
opreport prints its header information, you get something like the following:

  CPU: Intel Sandy Bridge microarchitecture, speed 2.401e+06 MHz (estimated)
  Counted int_misc events (Instruction decoder events) with a unit mask of 0x00
  (rat_stall_cycles Cycles Resource Allocation Table (RAT) external stall is
  sent to Instruction Decode Queue (IDQ) for this thread.) count 2000000

Notice that the unit mask value '0x00' is shown, even though the code actually
selects the default unit value of 0x40 for the int_misc event.

This patch fixes this issue.  It also partially addresses the issue with
named unit mask showing up as '0x00' in opreport, too (see oprofile bug
https://sourceforge.net/tracker/?func=detail&aid=3599234&group_id=16191&atid=116191
It's not a very good solution to the named unit mask issue, but it's a better
than nothing until we can come up with a final (better) solution.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>