Add Support for AMD Generic Performance Events
AMD generic performance events are a small set of events which are generally available across several
AMD processor families.  PERF has already provided supports for generic performance counters regardless
of the processor family.  This will allow operf to work as soon as PERF able to supports the performance
counters, and does not have to wait for the more complete family-specific events and unit_masks files
to be added to OProfile.

Signed-off-by: Suravee Suthikulpanit <suravee.suthikulpanit@amd.com>